requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
python-docx==0.8.11
PyPDF2==3.0.1
pydub==0.25.1
mutagen==1.47.0
librosa==0.10.1
soundfile==0.12.1
spacy==3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
numpy==1.26.4
pytest==8.0.2
black==24.2.0
flake8==7.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
ebooklib==0.18.0
openai==1.6.0
httpx==0.26.0
python-magic==0.4.27
SQLAlchemy==2.0.27
pydantic==2.5.2
typing-extensions==4.8.0
requests==2.31.0
Pillow==10.1.0
click==8.1.7
Jinja2==3.1.2
MarkupSafe==2.1.3
tqdm==4.66.1
tenacity==8.2.3 